Moved associated example fils to master:
It seems rather complex to have a workflow where example files like the
Dockerfile, and expected results have to be post-added using git foo on
the example branch itself.
Thus moving these to the master branch to be able to easily update them.

Updated the README accordingly.

Stil need to contribute the Jekins job DSL for the master demo runner
job, and common Jenkins job DSL for the example branch jobs.